Remove kernel_name_list and repetition guarding adn dead code from the opencl gen
Due to repetitions of the kernel names we used a list to avoid multiple
includes to the same file (to avoid ODR violation). With the previous
fix kernel names are unique so we dont need them.

Also removing some wierd code from the opencl generator which (afaik)
had no effect at all.